Check if new cover is avaible for chapter before downloading
Move some methods in mangadex/manga to keep the file sizes under control
