+++ main.go
@@ -46,6 +46,7 @@ func main() {
 	utils.Must(c.Provide(services.NotificationServiceProvider))
 	utils.Must(c.Provide(services.PreferenceServiceProvider))
 	utils.Must(c.Provide(services.ImageServiceProvider))
+	utils.Must(c.Provide(services.MetadataServiceProvider))
 	utils.Must(c.Provide(ApplicationProvider))
 
 	utils.Must(c.Invoke(services.RegisterSignalREndPoint))
